requirements.txt
streamlit
gspread==5.12.4
google-auth
docxtpl